% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEVDec.R
\name{FEVDec}
\alias{FEVDec}
\title{FEVDec}
\usage{
FEVDec(B_arr, Sig, lag, dec.type = "Generalized")
}
\arguments{
\item{B_arr}{arrays of autoregressive coeff}

\item{Sig}{var-cov matrix}

\item{lag}{forecast horizon}

\item{dec.type}{"Cholesky", "Spectral", "Generalized", "GeneralizedIRF" or "Generalized_Lanne". Default "Generalized".}
}
\value{
fevd: forecast error variance decompositoin matrix
}
\description{
Obtain Forecast Error variance Decomposition 
with Spectral, Cholesky or Generealized decomposition
}
\references{
Barbaglia, L., Croux, C., & Wilms, I. (2020). Volatility spillovers in commodity markets: A large t-vector autoregressive approach. Energy Economics, 85, 104555.
}
\author{
Luca Barbaglia \url{https://lucabarbaglia.github.io/}
}
