% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGrLasso.R
\name{JGrLasso}
\alias{JGrLasso}
\title{JGrLasso}
\usage{
JGrLasso(
  Y,
  X,
  P,
  n,
  J,
  type_lasso = "Lasso",
  maxiter.both = 50,
  tol.both = 0.01,
  lambda1.min = NULL,
  lambda1.max = NULL,
  lambda1.steps = NULL,
  gamma1.min = NULL,
  gamma1.max = NULL,
  gamma1.steps = NULL,
  lambda1.OPT = NULL,
  gamma1.OPT = NULL
)
}
\arguments{
\item{Y}{a vector of length J(N-P) cointaining the responses. Obtained with funxtion "stack_y". J: number of time series. N: time series length.}

\item{X}{a J(N-P)xPJ^2 matrix cointainig the regressor. Obtained with function "stack_Xbig".}

\item{P}{VAR order.}

\item{type_lasso}{type of lasso penalty. "Lasso" for standard lasso, "Group" for group lasso. Default is "Lasso".}

\item{tol.both}{tolerance gaussian lasso algorithm. Default is 0.01.}

\item{maxit.both}{maximum iterations for gaussian lasso algorithm. Default is 50.}

\item{lambda1_min}{minimum value of the regularization parameter on Beta. Default is NULL.}

\item{lambda1_max}{maximum value of the regularization parameter on Beta.  Default is NULL.}

\item{lambda1_steps}{number of steps in the lambda grid.  Default is NULL.}

\item{gamma1_min}{minimum value of the regularization parameter on Omega.  Default is NULL.}

\item{gamma1_max}{maximum value of the regularization parameter on Omega.  Default is NULL.}

\item{gamma1_steps}{number of steps in the gamma grid.  Default is NULL.}

\item{lambda1_OPT}{optimal value of the regularization parameter on Beta.  Default is NULL.}

\item{gamma1_OPT}{optimal value of the regularization parameter on Omega.  Default is NULL.}
}
\value{
A list containing two objects:
\item{"beta.new"}{a vector containing the estimated Beta.}
\item{"beta.arr"}{a JxJxP array containing the estimated Beta.}
\item{"omega.new"}{a JxJ matrix containing the estimated Omega.}
\item{"Obj_JGrL"}{objective function.}
\item{"iter"}{number of iterations.}
\item{"lambda"}{selected value of the regularization parameter on Beta.}
\item{"gamma"}{selected value of the regularization parameter on Omega.}
}
\description{
Function for the joint estimation of of beta and omega
}
\references{
Barbaglia, L., Croux, C., & Wilms, I. (2020). Volatility spillovers in commodity markets: A large t-vector autoregressive approach. Energy Economics, 85, 104555.
}
\author{
Luca Barbaglia \url{https://lucabarbaglia.github.io/}
}
