% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R
\docType{data}
\name{RV}
\alias{RV}
\title{Test data for Large t-VAR}
\format{
A matrix TxN of log-transformed realized volatilities for N=5 stocks and T=500 observations.
}
\usage{
data(RV)
}
\description{
A matrix TxN of log-transformed realized volatilities for N=5 stocks and T=500 observations.
}
\seealso{
\code{\link{Large.tVAR}} to fit a large t-VAR.
}
\keyword{datasets}
