% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARbyGLS.R
\name{VARbyGLS}
\alias{VARbyGLS}
\title{VARbyGLS}
\usage{
VARbyGLS(Y, p)
}
\arguments{
\item{Y}{mulitvariate time series.}

\item{p}{lag length.}
}
\value{
A list containing:
\item{"Beta"}{a list of betas.}
\item{"Sigma"}{var-cov matrix.}
\item{"Beta.vec"}{vector of betas [b^1_11,b^2_11,..,b^P_11,b^1_21,b^2_21,..,b^P_21,.., b^P_JJ].}
}
\description{
Function for the classic unpenalized VAR esitmation, Generalized Least Squares
}
