% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.vs.R
\name{network.vs}
\alias{network.vs}
\title{network.vs}
\usage{
network.vs(spill, main.label = NULL, names = NULL, ww = 5, colour = "#619CFF")
}
\arguments{
\item{spill}{matrix of spillover effects. Dimension JxJ, where J is the number of time series.}

\item{main.label}{network names.  Default NULL.}

\item{names}{node names. Default is the column names of beta.}

\item{ww}{weight of the thickness of the arrows}

\item{colour}{arrows' colour. Default "#619CFF".}
}
\value{
a network of spillover effects consisting of J nodes
}
\description{
Function to obtain network of volatility spillovers
}
